refactor: Product를 JPA entity로 변경, Repository 적용
