+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -6,15 +6,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.product.Product;
 
 import static javax.persistence.CascadeType.PERSIST;
 import static javax.persistence.CascadeType.REMOVE;
