+++ src/main/java/kitchenpos/domain/product/Product.java
@@ -1,15 +1,23 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
 
 import java.math.BigDecimal;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 
 import static java.util.Objects.isNull;
+import static javax.persistence.GenerationType.IDENTITY;
 
+@Entity
 public class Product {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
     private Long id;
     private String name;
     private BigDecimal price;
 
-    public Product() {
+    protected Product() {
     }
 
     public Product(final String name, final BigDecimal price) {
