+++ src/main/java/kitchenpos/domain/product/ProductRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.domain.product;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+}
