+++ src/test/java/kitchenpos/domain/MenuTest.java
@@ -3,6 +3,7 @@ package kitchenpos.domain;
 import java.math.BigDecimal;
 import java.util.Map;
 import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.product.Product;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
